requirements.txt
# Core AI Inference
litellm>=1.35.0
langchain-groq>=0.1.0

# Base LLM dependencies
openai>=1.14.0
pydantic>=2.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Evaluations
pytest>=8.0.0
